install react-quill and add the quill editor to the directions input. rework the controllers to new components and some small bug fixs
