+++ package.json
@@ -24,6 +24,7 @@
     "@trpc/next": "^10.26.0",
     "@trpc/react-query": "^10.26.0",
     "@trpc/server": "^10.26.0",
+    "@types/quill": "^2.0.10",
     "@vercel/analytics": "^1.0.1",
     "classnames": "^2.3.2",
     "next": "^13.4.2",
@@ -33,6 +34,7 @@
     "react-dom": "18.2.0",
     "react-hook-form": "^7.45.1",
     "react-icons": "^4.10.1",
+    "react-quill": "^2.0.0",
     "react-select": "^5.7.3",
     "superjson": "1.12.2",
     "uuidv4": "^6.2.13",
